chore(network): move db executor to a folder and separate dummy
