+++ crates/papyrus_network/src/db_executor/dummy_executor.rs
@@ -6,21 +6,9 @@ use std::task::Poll;
 use futures::Stream;
 use starknet_api::block::{BlockHeader, BlockSignature};
 
+use super::{DBExecutor, Data, QueryId};
 use crate::messages::protobuf;
 use crate::BlockQuery;
-
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub struct QueryId(pub usize);
-
-pub enum Data {
-    BlockHeaderAndSignature { header: BlockHeader, signature: BlockSignature },
-    Fin,
-}
-
-pub(crate) trait DBExecutor: Stream<Item = (QueryId, Data)> + Unpin {
-    // TODO: add writer functionality
-    fn register_query(&mut self, query: BlockQuery) -> QueryId;
-}
 
 pub struct DummyDBExecutor {
     _data: Vec<protobuf::BlockHeadersResponse>,
